Add better syntax highlighting for code blocks:
Use highlight.js for generating colored static HTML for use in code blocks.
Highlight.js is really cool: it has zero dependencies, works with a
bunch of different languages, has a bunch of styles you can apply, and
the generated HTML is relatively small.

Also use absolute naming for the stylesheets, which simplifies the template
generation a bit. This means that you wont be able to go to a file directly
via "file:///" anymore, but that isnt a huge deal.