Redesign markdown parser:
The old parser was built with the best of intentions, but had some major flaws:
Everything was parsed on a line-by-line basis, which made ignoring code in code
blocks and such impossible. It also was a lot harder to reason about, and the
HTML generation was tangled up in "AST" generator. This new version instead
uses a lot of small "phases" or "passes", which allows for easily changing how
nodes are grouped together.