Rename blogs, fix blog dates not being rendered correctly:
For starters, the dates where not sorted. Secondly, the dates for renamed files
only reached back 1 level, meaning renaming a file 2 times would not preserve the
original date. To fix this, we recursively backtrack renamed files, finding the
original commit for that file, and get it's date. There might be a better way to
do this, but this works for now.