Use match statement in more places, cleanup CodeblockNode:
I figured it would be good to try and replace some of these isinstance if
checks with match statements. They are great for the most part, but it would
be nice to not have to use keyword arguments to pull feilds out. There seems
to be a conflict with kw_only and match_arg dataclasses, so I will need to
look into that more.