Fix poor mobile rendering:
Due to the viewport settings for mobile browsers, the pages where rendering as
zoomed in by default, and the best way to counteract this was to make the <main>
element fill up the entire width, and scale the text to match it. Doesnt exactly
match the font style of the desktop version, but it still is around 80 chars
wide, so that is fine by me.