Add Open Graph titles for blog posts:
I could also add a description by pulling the first N characters from the blog,
but I was feeling lazy and felt that this would be a good improvement. It
would also be cool to have an image of sorts, but it would either have to be a
generic image I put on everything, or some autogenerated one for each. Either
way it is simpler to just not include an image or description for now.