Fix indentation on `ol` and `ul` elements:
The last commit didn't do what I wanted, so I redid it to make it display how I
wanted. Basically, I wanted to have the numbers left-aligned, and if the list
got over a certain size (ie, 10 or 100 elements) the margin would be the same
across each list element. I could not for the life of me get it to work with
CSS, so I resorted to statically inserting the data in there. I might try to
make it work with pure CSS later, but this will suffice for now.