Rewrite HTML generator to use the visitor pattern:
This makes the escaping, expanding, and generation of HTML much simpler and
cohesive. Currently there is no tree-like structure to the nodes, but later on
I will be making dedicated nodes for things like inline bold and italics,
nested bullet lists and more.

Another thing I have yet to do is split this into its own package, since I am
needing to use it in more places. The code isn't generic enough for me to split
it up yet, but in a few commits I hope to have it ready to relocate.